+++ mining/src/mempool/mod.rs
@@ -13,9 +13,9 @@ use std::rc::Rc;
 pub(crate) mod check_transaction_standard;
 pub mod config;
 pub mod errors;
-pub(crate) mod fill_inputs_and_get_missing_parents;
 pub(crate) mod handle_new_block_transactions;
 mod model;
+pub(crate) mod populate_entries_and_try_validate;
 pub(crate) mod remove_transaction;
 pub(crate) mod revalidate_high_priority_transactions;
 pub(crate) mod validate_and_insert_transaction;
